Initial commit of many refactoring changes
