+++ go.mod
@@ -1,5 +1,7 @@
 module github.com/denverquane/amongusdiscord
 
+go 1.15
+
 // +heroku goVersion go1.15
 
 require (
