+++ main.go
@@ -7,7 +7,6 @@ import (
 	"os"
 	"os/signal"
 	"strconv"
-	"strings"
 	"syscall"
 	"time"
 
@@ -79,26 +78,26 @@ func discordMainWrapper() error {
 	if err != nil {
 		numShards = 1
 	}
-	ports := make([]string, numShards)
-	tempPort := strings.ReplaceAll(os.Getenv("PORT"), " ", "")
-	portStrings := strings.Split(tempPort, ",")
-	if len(ports) == 0 || len(tempPort) == 0 {
-		num, err := strconv.Atoi(tempPort)
 
-		if err != nil || num < 1024 || num > 65535 {
-			log.Printf("[Info] Invalid or no particular PORT (range [1024-65535]) provided. Defaulting to %s\n", DefaultPort)
-			ports[0] = DefaultPort
-		}
-	} else if len(portStrings) == numShards {
-		for i := 0; i < numShards; i++ {
-			num, err := strconv.Atoi(portStrings[i])
-			if err != nil || num < 0 || num > 65535 {
-				return errors.New("invalid or no particular PORT (range [0-65535]) provided")
-			}
-			ports[i] = portStrings[i]
-		}
+	shardIDStr := os.Getenv("SHARD_ID")
+	shardID, err := strconv.Atoi(shardIDStr)
+	if shardID >= numShards {
+		return errors.New("you specified a shardID higher than or equal to the total number of shards")
+	}
+	if err != nil {
+		shardID = 0
+	}
+
+	portStr := os.Getenv("PORT")
+	port := ""
+
+	num, err := strconv.Atoi(portStr)
+
+	if err != nil || num < 1024 || num > 65535 {
+		log.Printf("[Info] Invalid or no particular PORT (range [1024-65535]) provided. Defaulting to %s\n", DefaultPort)
+		port = DefaultPort
 	} else {
-		return errors.New("the number of shards does not match the number of ports provided")
+		port = portStr
 	}
 
 	url := os.Getenv("SERVER_URL")
@@ -153,18 +152,12 @@ func discordMainWrapper() error {
 	sc := make(chan os.Signal, 1)
 	signal.Notify(sc, syscall.SIGINT, syscall.SIGTERM, os.Interrupt, os.Kill)
 
-	bots := make([]*discord.Bot, numShards)
+	bot := discord.MakeAndStartBot(VERSION, discordToken, discordToken2, url, port, extPort, emojiGuildID, numShards, shardID, storageClient)
 
-	for i := 0; i < numShards; i++ {
-		bots[i] = discord.MakeAndStartBot(VERSION, discordToken, discordToken2, url, ports[i], extPort, emojiGuildID, numShards, i, storageClient)
-	}
-
-	go discord.MessagesServer("5000", bots)
+	go discord.MessagesServer("5000", bot)
 
 	<-sc
-	for i := 0; i < numShards; i++ {
-		bots[i].Close()
-	}
+	bot.Close()
 	storageClient.Close()
 	return nil
 }
