+++ storage/guildStats.go
@@ -1,0 +1,4 @@
+package storage
+
+type GuildStats struct {
+}
